% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_required_args}
\alias{handle_required_args}
\title{Checked required arg}
\usage{
handle_required_args(required_arg)
}
\arguments{
\item{required_arg}{Vector of required arguments.}
}
\description{
Function that handle missing arguments. It checks if there are required arguments that are missng and
print a message. It is supposed to be used inside a function.
}
